Header, Flexbox, Navbar
Amendments made to nav bar to make it better suited to a responsive
design.  A menu button appears when the size of the screen is smaller,
hiding the menu options.

Within the accordion sections, a image section of flex boxes has been
added. This is fully responsive.

The header title is now centrally aligned.